3.15大疆电机和达妙电机经典CAN均已适配新CAN类，达妙电机尚不适配FDCAN，原因有待查明

+++ SkyWalker/BSP/CAN/bsp_can.cpp
@@ -127,7 +127,7 @@ void BSP_CAN::Filter_Config(uint32_t id_type, uint32_t filter_index, uint32_t fi
  *
  * @return uint8_t 执行状态
  */
-uint8_t BSP_CAN::Send_Data(uint16_t ID, uint8_t *Data, uint16_t Length)
+uint8_t BSP_CAN::Send_Data(uint16_t ID, uint8_t *Data, uint32_t Length)
 {
     FDCAN_TxHeaderTypeDef tx_header;
